[feat] Implemented old way of defining the floor into the new System
